fix: Fix issue where rollup keeps requesting indefinitely by submitting only BOB after 3 seconds even if TOB hasn’t arrived at SBB

+++ eth/ethconfig/config_zkevm.go
@@ -120,7 +120,7 @@ type Zk struct {
 	SeedNodeUrl                string
 	SequencerPrivateKey        string
 	LighthouseUrl              string
-	SlotTime                   int
+	SlotTime                   uint64
 	SbbUrl                     string
 }
